requirements.txt
langchain_openai
langchain_google_genai
browser_use
requests
python-dotenv
pydantic